pyproject.toml
[tool.poetry]
name = "DeefBot"
version = "0.1.0"
description = "Some Discord bot that can do things, created in Python"
authors = ["Dave Roverts <me@daveroverts.nl>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/daveroverts/DeefBot"

[tool.poetry.dependencies]
python = "^3.7"
discord-py = "^1.3.4"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
